+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Please mark all change in change log and use the ticket from JIRA.
 - \#602 - Optimizer specify wrong gpu_id
 - \#606 - No log generated during building index with CPU
 - \#631 - FAISS isn't compiled with O3 option
+- \#640 - Read all files as READONLY inverted list
 
 ## Feature
 - \#12 - Pure CPU version for Milvus